PMK-2352: Addressed review comments and enabled password change functionality.
