+++ src/app/core/components/UserMenu.js
@@ -5,6 +5,7 @@ import { withStyles } from '@material-ui/styles'
 import { withRouter } from 'react-router'
 import { withAppContext } from 'core/providers/AppProvider'
 import { logoutUrl } from 'app/constants'
+import ChangePasswordModal from './ChangePasswordModal'
 
 const styles = theme => ({
   avatar: {
@@ -24,20 +25,23 @@ const styles = theme => ({
 @withStyles(styles)
 @withRouter
 class UserMenu extends React.PureComponent {
-  state = { anchorEl: null }
+  state = { anchorEl: null, showChangePasswordModal: false }
   handleClick = event => this.setState({ anchorEl: event.currentTarget })
   handleClose = () => this.setState({ anchorEl: null })
   logout = () => this.props.history.push(logoutUrl)
 
-  handleChangePassword = () => {
-    this.props.setContext({ isChangePassword: true })
-    this.handleClose()
-  }
+  handleChangePassword = () => this.setState({ showChangePasswordModal: true, anchorEl: null })
+
+  handleCancelChangePassword = () => this.setState({ showChangePasswordModal: false })
 
   render () {
     const { classes, className, session } = this.props
-    const { anchorEl } = this.state
+    const { anchorEl, showChangePasswordModal } = this.state
     const username = session.username || '?'
+
+    if (showChangePasswordModal) {
+      return <ChangePasswordModal onCancel={this.handleCancelChangePassword} />
+    }
 
     return (
       <div className={`${classes.avatar} ${className}`}>
@@ -50,7 +54,7 @@ class UserMenu extends React.PureComponent {
           getContentAnchorEl={null}
           anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
         >
-          { false && <MenuItem onClick={this.handleChangePassword}>Change Password</MenuItem> }
+          <MenuItem onClick={this.handleChangePassword}>Change Password</MenuItem>
           <MenuItem onClick={this.logout}>Sign Out</MenuItem>
         </Menu>
       </div>
