+++ src/app/core/public/LoginPage.js
@@ -15,7 +15,6 @@ import { forgotPasswordUrl } from 'app/constants.js'
 import { pathJoin } from 'utils/misc'
 import { imageUrlRoot, dashboardUrl } from 'app/constants'
 import moment from 'moment'
-import ToastContainer from 'core/components/toasts/ToastContainer'
 
 const styles = theme => ({
   root: {
@@ -169,28 +168,16 @@ export class LoginPage extends React.PureComponent {
   }
 
   render () {
-    const { classes, history: { location: { toastData } } } = this.props
+    const { classes } = this.props
     const { loginFailed, loading } = this.state
 
     return (
       <div className="login-page">
-        {toastData && (
-          <ToastContainer
-            toasts={[
-              {
-                id: toastData.id,
-                isOpen: toastData.isOpen,
-                variant: toastData.variant,
-                text: toastData.text,
-              },
-            ]}
-          />
-        )}
         <Grid container justify="center" className={classes.root}>
           <Grid item md={4} lg={3}>
             <Paper className={classes.paper}>
               <img
-                alt="Platform 9"
+                alt="Platform9"
                 src={pathJoin(imageUrlRoot, 'logo-color.png')}
                 className={classes.img}
               />
